Restructure to `src` and `build` dirs, major renaming
BREAKING CHANGE:

- Default click action options renamed:
    - `"close"`    -> `"windowlist-close"`
    - `"minimize"` -> `"windowlist-minimize"`
    - `"raise"`    -> `"windowlist-raise"`

Unfortunately breaks existing `config.toml` files that use the old
default values.
