+++ config-default.toml
@@ -10,8 +10,8 @@ name_padding = 1
 separator_string = "·"
 empty_desktop_string = ""
 
-active_window_left_click          = "minimize"
-active_window_right_click         = "close"
+active_window_left_click          = "windowlist-minimize"
+active_window_right_click         = "windowlist-close"
 active_window_middle_click        = "none"
 active_window_left_double_click   = "none"
 active_window_right_double_click  = "none"
@@ -19,8 +19,8 @@ active_window_middle_double_click = "none"
 active_window_scroll_up           = "none"
 active_window_scroll_down         = "none"
 
-inactive_window_left_click          = "raise"
-inactive_window_right_click         = "close"
+inactive_window_left_click          = "windowlist-raise"
+inactive_window_right_click         = "windowlist-close"
 inactive_window_middle_click        = "none"
 inactive_window_left_double_click   = "none"
 inactive_window_right_double_click  = "none"
