+++ src/click-actions/windowlist-raise.c
@@ -2,8 +2,8 @@
 #include <stdlib.h>
 #include <X11/Xlib.h>
 #include <X11/Xatom.h>
-#include "common.h"
-#include "../../windowlist.h" // get_desktop_id()
+#include "actions-common.h"
+#include "../xlib-utils.h" // get_desktop_id()
 
 int main(int argc, char* argv[]) {
     // Must take a window id as first argument
