+++ src/xlib-utils.c
@@ -18,7 +18,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <X11/Xatom.h>
-#include "windowlist.h"
+#include "xlib-utils.h"
 
 #define MAX_PROPERTY_VALUE_LEN 4096
