+++ src/xlib-utils.h
